requirements.txt
Flask==3.0.3
openai==1.41.0
praw==7.7.1
python-dotenv==1.0.0